Make the plugin public and separate it from the implementations
