+++ tests/unit/lms/services/grouping/_plugin/canvas_test.py
@@ -3,8 +3,8 @@ from unittest.mock import sentinel
 import pytest
 
 from lms.services import CanvasAPIError
-from lms.services.grouping._plugin import GroupError
 from lms.services.grouping._plugin.canvas import CanvasGroupingPlugin, ErrorCodes
+from lms.services.grouping.plugin import GroupError
 from tests import factories
